rewrote the whole file to include the sel function.
the sel function in simple words is a for loop. it allows the user to perform one category of operation until user quits. (like basic, adv)
the user types bcalc sel basic to select the basic category and to enable all its operations. the user can then proceed to add, sub, or any function. one can press q to quit.